+++ tests/integration/cloudformation/api/test_stacks.snapshot.json
@@ -413,67 +413,6 @@
       "deletion": true
     }
   },
-  "tests/integration/cloudformation/api/test_stacks.py::test_drift_detection_on_lambda": {
-    "recorded-date": "26-10-2022, 10:51:32",
-    "recorded-content": {
-      "drift_detection": {
-        "StackResourceDrift": {
-          "ActualProperties": {
-            "Description": "different description",
-            "Environment": {
-              "Variables": {
-                "ENDPOINT_URL": "localhost.localstack.cloud"
-              }
-            },
-            "Handler": "index.handler",
-            "Role": "arn:aws:iam::111111111111:role/<resource:1>",
-            "Runtime": "python3.8"
-          },
-          "ExpectedProperties": {
-            "Description": "function to test lambda function url",
-            "Environment": {
-              "Variables": {
-                "ENDPOINT_URL": "aws.amazon.com"
-              }
-            },
-            "Handler": "index.handler",
-            "Role": "arn:aws:iam::111111111111:role/<resource:1>",
-            "Runtime": "python3.9"
-          },
-          "LogicalResourceId": "Function76856677",
-          "PhysicalResourceId": "stack-9899e521-Function76856677-KuUf2SS1Sr7H",
-          "PropertyDifferences": [
-            {
-              "ActualValue": "different description",
-              "DifferenceType": "NOT_EQUAL",
-              "ExpectedValue": "function to test lambda function url",
-              "PropertyPath": "/Description"
-            },
-            {
-              "ActualValue": "localhost.localstack.cloud",
-              "DifferenceType": "NOT_EQUAL",
-              "ExpectedValue": "aws.amazon.com",
-              "PropertyPath": "/Environment/Variables/ENDPOINT_URL"
-            },
-            {
-              "ActualValue": "python3.8",
-              "DifferenceType": "NOT_EQUAL",
-              "ExpectedValue": "python3.9",
-              "PropertyPath": "/Runtime"
-            }
-          ],
-          "ResourceType": "AWS::Lambda::Function",
-          "StackId": "arn:aws:cloudformation:<region>:111111111111:stack/stack-9899e521/<resource:2>",
-          "StackResourceDriftStatus": "MODIFIED",
-          "Timestamp": "timestamp"
-        },
-        "ResponseMetadata": {
-          "HTTPHeaders": {},
-          "HTTPStatusCode": 200
-        }
-      }
-    }
-  },
   "tests/integration/cloudformation/api/test_stacks.py::test_prevent_stack_deletion_and_update": {
     "recorded-date": "28-10-2022, 11:41:53",
     "recorded-content": {}
@@ -485,5 +424,21 @@
   "tests/integration/cloudformation/api/test_stacks.py::test_prevent_resource_deletion": {
     "recorded-date": "28-10-2022, 12:29:11",
     "recorded-content": {}
+  },
+  "tests/integration/cloudformation/api/test_stacks.py::test_linting_error_during_creation": {
+    "recorded-date": "11-11-2022, 08:10:14",
+    "recorded-content": {
+      "error": {
+        "Error": {
+          "Code": "ValidationError",
+          "Message": "Template format error: Any Resources member must be an object.",
+          "Type": "Sender"
+        },
+        "ResponseMetadata": {
+          "HTTPHeaders": {},
+          "HTTPStatusCode": 400
+        }
+      }
+    }
   }
 }
